requirements.txt
disnake==2.9.2
python-dotenv==1.0.0
psutil==5.9.6
asyncio==3.4.3
aiohttp==3.8.6
colorama==0.4.6
